requirements.txt
torch
transformers
sentence-transformers
opencv-python
pillow
sounddevice
faster-whisper
anthropic
python-dotenv
faiss-cpu
PyQt5
PyYAML
ultralytics
piper-tts==1.2.0
piper-phonemize==1.1.0
